Fix issue syncing deleted directories
This fixes an issue where a directory in the remote directory containing
only files marked for deletion would be synced back to the local
directory as an empty directory. This means that a user who deleted a
directory locally would sometimes find an empty directory in its place
after a sync.