Disallow commenting out options in config file
The original motivation for allowing certain config options to be
commented out was to make it easy to return options to their default
values and to remind the user of which options they had modified.
Disallowing this means that the default values can change in future
versions without affecting existing profiles. It also means that all
default values are now stored in only one place. Users can still find
the original values in the master template config file.