Allow commenting out options in the config file
Storing default values in the code as well as in the config file and
allowing certain options to be commented out makes it easy to add
additional options in the future without requiring users to update their
config file. These options are not commented out by default so that
changes to default values don't affect existing users.