Fix issue handling conflicts
This commit fixes an issue that would cause new files created in the
process of resolving conflicts to not be added to the database.

As of this commit, new files synced from the remote directory do not
have their priority inflated. Only new files created locally have their
priority inflated.