Remove ability to mount a remote filesystem
This commit removes all code for mounting a filesystem over SSH.
Instead, the remote directory must be accessible in the local
filesystem. A network filesystem can still be used, but it is up to the
user to mount it. The purpose of this change is to make the program
agnostic to the connection method being used so that NFS and SMB can
also be used.