Remove rsync as a dependency
Rsync was originally used for transferring files because of its
delta-transfer algorithm that reduces the amount of data sent over the
network. However, this algorithm isn't used when transferring files
locally, which includes network filesystems. In this case, the algorithm
is actually slower than transferring files normally. This commit
replaces rsync with Python code and removes it as a dependency. This is
also a step toward making the program cross-platform in the future.